Add lazy initialization in MariaDBStore
- Introduced `lazy_init` parameter to defer initialization of the vector store.
- Added unit tests for lazy initialization scenarios.

Motivation:
The primary motivation for this patch is that,
in a client-server application (e.g., RAG), the user might not know the embedding_length in advance.
Do determine the length, we want to avoid making an initial "fake" request to the embedding endpoint
on behalf of the user, which could incur unnecessary costs.

Instead, we now support lazy initialization of the store.
This allows us to wait until the first user request to store a document.
The API can simply call add_texts on a MariaDBStore instance when needed.